requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.73.0
h11==0.13.0
httptools==0.3.0
idna==3.3
motor==2.5.1
pydantic==1.9.0
pymongo==3.12.3
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.4
watchgod==0.8.2
websockets==10.2
PyJWT==1.7.1
python-decouple==3.3
email-validator==1.1.2
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.73.0
h11==0.13.0
httptools==0.3.0
idna==3.3
motor==2.5.1
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.4
watchgod==0.8.2
websockets==10.2
